Add more multipleOf tests
There were already some tests for fractional multipleOf values,
but while writing an implementation I discovered that those were
not enough to catch a certain kind of poor implementations.

As JSON encodes arbitrary numbers in decimal form, the correct
implementation should either use lossless decimal presentation, or at
least operate in "shortest decimal which is presented as A is multiple
of shortest decimal which is presented as B" to be conforming.